Moved HLMatrixBitcastLowerPass to its own file. Further implemented matrix subscript lowering.
