+++ lib/HLSL/CMakeLists.txt
@@ -22,6 +22,7 @@ add_llvm_library(LLVMHLSL
   DxilExportMap.cpp
   DxilValidation.cpp
   DxcOptimizer.cpp
+  HLMatrixBitcastLowerPass.cpp
   HLMatrixLowerPass.cpp
   HLMatrixType.cpp
   HLModule.cpp
